added segmentio kafka-go producer and consumer
