+++ kafka-go/go.mod
@@ -1,0 +1,12 @@
+module example.com/kafka-go
+
+go 1.18
+
+require (
+	github.com/segmentio/kafka-go v0.4.30
+)
+
+require (
+	github.com/klauspost/compress v1.14.2 // indirect
+	github.com/pierrec/lz4/v4 v4.1.14 // indirect
+)
